requirements.txt
tensorflow
opencv-python
pillow
imagehash
tensorflow-serving-api
google-cloud-vision
matplotlib
sklearn